Small change to setting of HTTP_HOST in tests.InstanceClient.
Instead of hard-coding HTTP_HOST in just get and post requests
on InstanceClient, set HTTP_HOST in the client's defaults if
it's not set there explicitly.

This has a few beneficial effects:

1) It works for all kinds of request, not just GET and POST.
2) Individual requests can now override this by setting the kwarg
(before this was just discarded)
3) If you wanted for some reason to subclass this client and set
a different HTTP_HOST, you now can. I can't see why you'd want to
do this third one at the moment as the only thing that makes
InstanceClient different is the HTTP_HOST...